+++ picasso/src/main/java/com/squareup/picasso/LruCache.java
@@ -15,15 +15,10 @@
  */
 package com.squareup.picasso;
 
-import android.app.ActivityManager;
 import android.content.Context;
 import android.graphics.Bitmap;
-import android.os.Build;
 import java.util.LinkedHashMap;
 import java.util.Map;
-
-import static android.content.Context.ACTIVITY_SERVICE;
-import static android.content.pm.ApplicationInfo.FLAG_LARGE_HEAP;
 
 /** A memory cache which uses a least-recently used eviction policy. */
 public class LruCache implements Cache {
@@ -38,7 +33,7 @@ public class LruCache implements Cache {
 
   /** Create a cache using an appropriate portion of the available RAM as the maximum size. */
   public LruCache(Context context) {
-    this(calculateMaxSize(context));
+    this(Utils.calculateMemoryCacheSize(context));
   }
 
   /** Create a cache with a given maximum size in bytes. */
@@ -143,21 +138,5 @@ public class LruCache implements Cache {
   /** Returns the number of values that have been evicted. */
   public final synchronized int evictionCount() {
     return evictionCount;
-  }
-
-  private static int calculateMaxSize(Context context) {
-    ActivityManager am = (ActivityManager) context.getSystemService(ACTIVITY_SERVICE);
-    boolean largeHeap = (context.getApplicationInfo().flags & FLAG_LARGE_HEAP) != 0;
-    int memoryClass = am.getMemoryClass();
-    if (largeHeap && Build.VERSION.SDK_INT <= Build.VERSION_CODES.HONEYCOMB) {
-      memoryClass = ActivityManagerHoneycomb.getLargeMemoryClass(am);
-    }
-    return 1024 * 1024 * memoryClass / 6;
-  }
-
-  private static class ActivityManagerHoneycomb {
-    static int getLargeMemoryClass(ActivityManager activityManager) {
-      return activityManager.getLargeMemoryClass();
-    }
   }
 }
