+++ picasso/src/main/java/com/squareup/picasso/Picasso.java
@@ -491,7 +491,22 @@ public class Picasso {
     return result;
   }
 
-  /** The global default {@link Picasso} instance. */
+  /**
+   * The global default {@link Picasso} instance.
+   * <p>
+   * This instance is automatically initialized with defaults that are suitable to most
+   * implementations.
+   * <ul>
+   * <li>LRU memory cache of 15% the available application RAM up to 20MB</li>
+   * <li>Disk cache of 2% storage space up to 50MB but no less than 5MB. (Note: this is only
+   * available on API 14+ <em>or</em> if you are using a standalone library that provides a disk
+   * cache on all API levels like OkHttp)</li>
+   * <li>Three download threads for disk and network access.</li>
+   * </ul>
+   * <p>
+   * If these settings do not meet the requirements of your application you can construct your own
+   * instance with full control over the configuration by using {@link Picasso.Builder}.
+   */
   public static Picasso with(Context context) {
     if (singleton == null) {
       singleton = new Builder(context).build();
