+++ picasso/src/main/java/com/squareup/picasso/Utils.java
@@ -1,18 +1,24 @@
 package com.squareup.picasso;
 
+import android.app.ActivityManager;
 import android.content.ContentResolver;
 import android.content.Context;
 import android.database.Cursor;
 import android.graphics.Bitmap;
 import android.media.ExifInterface;
 import android.net.Uri;
+import android.os.Build;
 import android.os.Looper;
 import android.os.Process;
+import android.os.StatFs;
 import android.provider.MediaStore;
+import java.io.File;
 import java.io.IOException;
 import java.util.List;
 import java.util.concurrent.ThreadFactory;
 
+import static android.content.Context.ACTIVITY_SERVICE;
+import static android.content.pm.ApplicationInfo.FLAG_LARGE_HEAP;
 import static android.media.ExifInterface.ORIENTATION_NORMAL;
 import static android.media.ExifInterface.ORIENTATION_ROTATE_180;
 import static android.media.ExifInterface.ORIENTATION_ROTATE_270;
@@ -26,6 +32,9 @@ final class Utils {
   static final String THREAD_PREFIX = "Picasso-";
   static final String THREAD_IDLE_NAME = THREAD_PREFIX + "Idle";
   private static final int KEY_PADDING = 50; // Determined by exact science.
+  private static final int MIN_DISK_CACHE_SIZE = 5 * 1024 * 1024; // 5MB
+  private static final int MAX_DISK_CACHE_SIZE = 50 * 1024 * 1024; // 50MB
+  private static final int MAX_MEM_CACHE_SIZE = 20 * 1024 * 1024; // 20MB
   private static final String[] CONTENT_ORIENTATION = new String[] {
       MediaStore.Images.ImageColumns.ORIENTATION
   };
@@ -178,6 +187,34 @@ final class Utils {
       return OkHttpLoaderCreator.create(context);
     } catch (ClassNotFoundException e) {
       return new UrlConnectionLoader(context);
+    }
+  }
+
+  static int calculateDiskCacheSize(File dir) {
+    StatFs statFs = new StatFs(dir.getAbsolutePath());
+    int available = statFs.getBlockCount() * statFs.getBlockSize();
+    // Target 2% of the total space.
+    int size = available / 50;
+    // Bound inside min/max size for disk cache.
+    return Math.max(Math.min(size, MAX_DISK_CACHE_SIZE), MIN_DISK_CACHE_SIZE);
+  }
+
+  static int calculateMemoryCacheSize(Context context) {
+    ActivityManager am = (ActivityManager) context.getSystemService(ACTIVITY_SERVICE);
+    boolean largeHeap = (context.getApplicationInfo().flags & FLAG_LARGE_HEAP) != 0;
+    int memoryClass = am.getMemoryClass();
+    if (largeHeap && Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB) {
+      memoryClass = ActivityManagerHoneycomb.getLargeMemoryClass(am);
+    }
+    // Target 15% of the available RAM.
+    int size = 1024 * 1024 * memoryClass / 7;
+    // Bound to max size for mem cache.
+    return Math.min(size, MAX_MEM_CACHE_SIZE);
+  }
+
+  private static class ActivityManagerHoneycomb {
+    static int getLargeMemoryClass(ActivityManager activityManager) {
+      return activityManager.getLargeMemoryClass();
     }
   }
